Go back to splitting EngineConfig and Engine
...but this time, we call Engine, what previously was the EngineConfig
and call SystemOfQuantities what earlier was the Engine.

Quantities still have the Engine as template parameter, which means
their parameter changed from what we now call the SystemOfQuantities to
what we previously called the EngineConfig.

This step was necessary, because I was right when I thought that we'd
get template instanciation loops and was wrong when I thought we
wouldn't.

Using the EngineConfig as the template parameter reduces the risk of
getting more recursions.