Simplify template structure of Quantity
So far, quantities were templated with an engine and a list of
DimensionComponent types, e.g. Quantity<SiEngine,
DimensionComponent<base_dimensions::Mass, 1>, DimensionComponent<base_dimensions::Length, 1>,
DimensionComponent<base_dimensions::Time, -2>>. The order was already
relevant, but base-dimensions with exponent 0 where left out. The
advantage of that is, that the base dimensions are universal (as long as
the id is unique).

Now, we template the quantities with an engine and a list of integers
that designate the exponents of the base dimension. The relation to the
base dimensions is encoded in the position and specified by the engine
(explicitly the engine's system of dimensions), thus the exponents of all
dimensions must always be listed.
This greatly simplifies the template logic involved in multiplications
and divisions. It drastically reduces the size length of the full type name.
Moreover it helps reduce the boilerplate code involved in creating custom
engines and finally makes the base code of quantityland2 completely
agnostic to the dimensions that exist (while before, it specified some
base_dimensions).